// net/js/enterprise-nx6-workspace/libs/commerce/basket-lib/src/api/model/basketItem.ts
/**
 * Basket HTTP API
 * The Basket Service HTTP API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface BasketItem {
    id?: string;
    productId?: string;
    productName?: string;
    unitPrice?: number;
    oldUnitPrice?: number;
    quantity?: number;
    pictureUrl?: string;
}